+++ django_generic_api/tests/API_fixtures.py
