+++ django_generic_api/tests/test_fetch.py
@@ -3,7 +3,7 @@ import pytest
 from rest_framework.test import APIClient
 from api_app.models import Customer
 import json
-from fixtures import (
+from API_fixtures import (
     api_client,
     fetch_data_1,
     view_perm_token,
